refactor: fold kv test files in a single one and add comment explaining the purpose of kv.go file
